test: keep goto split regression within test file budget
